+++ python/pyspark/cloudpickle/__init__.py
@@ -1,8 +1,18 @@
+from pyspark.cloudpickle import cloudpickle  # noqa
 from pyspark.cloudpickle.cloudpickle import *  # noqa
-from pyspark.cloudpickle.cloudpickle_fast import CloudPickler, dumps, dump  # noqa
 
-# Conform to the convention used by python serialization libraries, which
-# expose their Pickler subclass at top-level under the  "Pickler" name.
-Pickler = CloudPickler
+__doc__ = cloudpickle.__doc__
 
-__version__ = '2.2.1'
+__version__ = "3.0.0"
+
+__all__ = [  # noqa
+    "__version__",
+    "Pickler",
+    "CloudPickler",
+    "dumps",
+    "loads",
+    "dump",
+    "load",
+    "register_pickle_by_value",
+    "unregister_pickle_by_value",
+]
